+++ mex/extractors/synopse/main.py
@@ -1,9 +1,13 @@
 from itertools import chain, groupby, tee
+from typing import Any
 
 from mex.common.cli import entrypoint
 from mex.common.ldap.extract import get_merged_ids_by_query_string
 from mex.common.ldap.models.person import LDAPPersonWithQuery
-from mex.common.ldap.transform import transform_ldap_persons_with_query_to_mex_persons
+from mex.common.ldap.transform import (
+    transform_ldap_actors_to_mex_contact_points,
+    transform_ldap_persons_with_query_to_mex_persons,
+)
 from mex.common.models import (
     AccessPlatformMapping,
     ActivityMapping,
@@ -15,8 +19,8 @@ from mex.common.models import (
     ExtractedVariableGroup,
     ResourceMapping,
 )
-from mex.common.organigram.extract import get_unit_merged_ids_by_emails
 from mex.common.types import (
+    MergedContactPointIdentifier,
     MergedOrganizationalUnitIdentifier,
     MergedOrganizationIdentifier,
     MergedPersonIdentifier,
@@ -29,20 +33,22 @@ from mex.extractors.synopse.extract import (
     extract_projects,
     extract_study_data,
     extract_study_overviews,
+    extract_synopse_contact,
     extract_synopse_organizations,
     extract_synopse_project_contributors,
     extract_variables,
 )
-from mex.extractors.synopse.filter import filter_and_log_access_platforms
+from mex.extractors.synopse.filter import (
+    filter_and_log_access_platforms,
+    filter_and_log_synopse_variables,
+)
 from mex.extractors.synopse.models.project import SynopseProject
 from mex.extractors.synopse.models.study import SynopseStudy
 from mex.extractors.synopse.models.study_overview import SynopseStudyOverview
 from mex.extractors.synopse.models.variable import SynopseVariable
 from mex.extractors.synopse.transform import (
-    split_off_extended_data_use_variables,
     transform_overviews_to_resource_lookup,
-    transform_synopse_data_extended_data_use_to_mex_resources,
-    transform_synopse_data_regular_to_mex_resources,
+    transform_synopse_data_to_mex_resources,
     transform_synopse_projects_to_mex_activities,
     transform_synopse_studies_into_access_platforms,
     transform_synopse_variables_to_mex_variable_groups,
@@ -78,25 +84,21 @@ def synopse_study_overviews() -> list[SynopseStudyOverview]:
 
 
 @asset(group_name="synopse")
-def synopse_variables() -> list[SynopseVariable]:
+def synopse_variables(
+    extracted_primary_source_report_server: ExtractedPrimarySource,
+) -> list[SynopseVariable]:
     """Extract variables from Synopse."""
-    return list(extract_variables())
+    return filter_and_log_synopse_variables(
+        extract_variables(), extracted_primary_source_report_server
+    )
 
 
 @asset(group_name="synopse")
-def synopse_variables_extended_data_use_by_study_id(
+def synopse_variables_by_study_id(
     synopse_variables: list[SynopseVariable],
-    synopse_study_overviews: list[SynopseStudyOverview],
 ) -> dict[int, list[SynopseVariable]]:
-    """Convert Synopse data to synopse_variables_extended_data_use_by_study_id."""
-    (
-        _,
-        variables_extended_data_use,
-    ) = split_off_extended_data_use_variables(
-        synopse_variables, synopse_study_overviews
-    )
-
-    sorted_variables = sorted(variables_extended_data_use, key=lambda v: v.studie_id)
+    """Convert Synopse data to synopse_variables_by_study_id."""
+    sorted_variables = sorted(synopse_variables, key=lambda v: v.studie_id)
     return {
         studie_id: list(variables)
         for studie_id, variables in groupby(sorted_variables, key=lambda v: v.studie_id)
@@ -104,39 +106,12 @@ def synopse_variables_extended_data_use_by_study_id(
 
 
 @asset(group_name="synopse")
-def synopse_variables_regular_use_by_study_id(
+def synopse_variables_by_thema(
     synopse_variables: list[SynopseVariable],
-    synopse_study_overviews: list[SynopseStudyOverview],
-) -> dict[int, list[SynopseVariable]]:
-    """Convert Synopse data to synopse_variables_regular_use_by_study_id."""
-    (
-        variables_regular,
-        _,
-    ) = split_off_extended_data_use_variables(
-        synopse_variables, synopse_study_overviews
-    )
-    sorted_variables = sorted(variables_regular, key=lambda v: v.studie_id)
-    return {
-        studie_id: list(variables)
-        for studie_id, variables in groupby(sorted_variables, key=lambda v: v.studie_id)
-    }
-
-
-@asset(group_name="synopse")
-def synopse_variables_regular_use_by_thema(
-    synopse_variables: list[SynopseVariable],
-    synopse_study_overviews: list[SynopseStudyOverview],
 ) -> dict[str, list[SynopseVariable]]:
-    """Convert Synopse data to synopse_variables_regular_use_by_thema."""
-    (
-        variables_regular,
-        _,
-    ) = split_off_extended_data_use_variables(
-        synopse_variables, synopse_study_overviews
-    )
-
+    """Convert Synopse data to synopse_variables_by_thema."""
     sorted_variables = sorted(
-        variables_regular, key=lambda v: v.thema_und_fragebogenausschnitt
+        synopse_variables, key=lambda v: v.thema_und_fragebogenausschnitt
     )
     return {
         thema: list(variables)
@@ -144,14 +119,6 @@ def synopse_variables_regular_use_by_thema(
             sorted_variables, key=lambda v: v.thema_und_fragebogenausschnitt
         )
     }
-
-
-@asset(group_name="synopse")
-def unit_stable_target_ids_by_emails(
-    extracted_organizational_units: list[ExtractedOrganizationalUnit],
-) -> dict[str, MergedOrganizationalUnitIdentifier]:
-    """Group organizational units by their email addresses."""
-    return get_unit_merged_ids_by_emails(extracted_organizational_units)
 
 
 @asset(group_name="synopse")
@@ -189,70 +156,68 @@ def synopse_organization_ids_by_query_string(
 
 
 @asset(group_name="synopse")
+def synopse_resource() -> dict[str, Any]:
+    """Extract and transform synopse resource default values."""
+    settings = Settings.get()
+    return load_yaml(settings.synopse.mapping_path / "resource.yaml")
+
+
+@asset(group_name="synopse")
+def contact_merged_id_by_query_string(
+    synopse_activity: dict[str, Any],
+    synopse_resource: dict[str, Any],
+    extracted_primary_source_ldap: ExtractedPrimarySource,
+) -> dict[str, MergedContactPointIdentifier]:
+    """Get lookup of ldap functional accounts by email."""
+    synopse_contact = extract_synopse_contact(
+        ResourceMapping.model_validate(synopse_resource),
+        ActivityMapping.model_validate(synopse_activity),
+    )
+    return {
+        contact_point.email[0].lower(): contact_point.stableTargetId
+        for contact_point in transform_ldap_actors_to_mex_contact_points(
+            synopse_contact,
+            extracted_primary_source_ldap,
+        )
+    }
+
+
+@asset(group_name="synopse")
 def extracted_synopse_resource_stable_target_ids_by_synopse_id(
     synopse_projects: list[SynopseProject],
     synopse_studies: list[SynopseStudy],
     synopse_study_overviews: list[SynopseStudyOverview],
-    synopse_variables_extended_data_use_by_study_id: dict[int, list[SynopseVariable]],
-    synopse_variables_regular_use_by_study_id: dict[int, list[SynopseVariable]],
+    synopse_variables_by_study_id: dict[int, list[SynopseVariable]],
     unit_stable_target_ids_by_synonym: dict[str, MergedOrganizationalUnitIdentifier],
     extracted_synopse_access_platforms: list[ExtractedAccessPlatform],
     extracted_synopse_activities: list[ExtractedActivity],
     extracted_organization_rki: ExtractedOrganization,
     extracted_primary_source_report_server: ExtractedPrimarySource,
+    synopse_resource: dict[str, Any],
+    contact_merged_id_by_query_string: dict[str, MergedContactPointIdentifier],
 ) -> dict[str, list[MergedResourceIdentifier]]:
     """Get lookup from synopse_id to extracted resource stable target id.
 
     Also transforms Synopse data to extracted resources
     """
-    settings = Settings.get()
-    synopse_resource = ResourceMapping.model_validate(
-        load_yaml(settings.synopse.mapping_path / "resource.yaml"),
+    transformed_study_data_resources = transform_synopse_data_to_mex_resources(
+        synopse_studies,
+        synopse_projects,
+        synopse_variables_by_study_id,
+        extracted_synopse_activities,
+        extracted_synopse_access_platforms,
+        extracted_primary_source_report_server,
+        unit_stable_target_ids_by_synonym,
+        extracted_organization_rki,
+        ResourceMapping.model_validate(synopse_resource),
+        contact_merged_id_by_query_string,
     )
-    transformed_study_data_regular_resources = (
-        transform_synopse_data_regular_to_mex_resources(
-            synopse_studies,
-            synopse_projects,
-            synopse_variables_regular_use_by_study_id,
-            extracted_synopse_activities,
-            extracted_synopse_access_platforms,
-            extracted_primary_source_report_server,
-            unit_stable_target_ids_by_synonym,
-            extracted_organization_rki,
-            synopse_resource,
-        )
-    )
-    transformed_study_data_regular_resource_gens = tee(
-        transformed_study_data_regular_resources, 2
-    )
-    load(transformed_study_data_regular_resource_gens[0])
-    settings = Settings.get()
-    synopse_resource_extended_data_use = ResourceMapping.model_validate(
-        load_yaml(settings.synopse.mapping_path / "resource_extended-data-use.yaml")
-    )
-    transformed_study_data_resources_extended_data_use = (
-        transform_synopse_data_extended_data_use_to_mex_resources(
-            synopse_studies,
-            synopse_projects,
-            synopse_variables_extended_data_use_by_study_id,
-            extracted_synopse_activities,
-            extracted_synopse_access_platforms,
-            extracted_primary_source_report_server,
-            unit_stable_target_ids_by_synonym,
-            extracted_organization_rki,
-            synopse_resource_extended_data_use,
-        )
-    )
-    transformed_study_data_resource_extended_data_use_gens = tee(
-        transformed_study_data_resources_extended_data_use, 2
-    )
-    load(transformed_study_data_resource_extended_data_use_gens[0])
-
+    transformed_study_data_resource_gens = tee(transformed_study_data_resources, 2)
+    load(transformed_study_data_resource_gens[0])
     return transform_overviews_to_resource_lookup(
         synopse_study_overviews,
         chain(
-            transformed_study_data_regular_resource_gens[1],
-            transformed_study_data_resource_extended_data_use_gens[1],
+            transformed_study_data_resource_gens[1],
         ),
     )
 
@@ -284,30 +249,34 @@ def extracted_synopse_access_platforms(
 
 
 @asset(group_name="synopse")
+def synopse_activity() -> dict[str, Any]:
+    """Extract and transform synopse activity default values."""
+    settings = Settings.get()
+    return load_yaml(settings.synopse.mapping_path / "activity.yaml")
+
+
+@asset(group_name="synopse")
 def extracted_synopse_activities(
     synopse_projects: list[SynopseProject],
     extracted_primary_source_report_server: ExtractedPrimarySource,
-    unit_stable_target_ids_by_emails: dict[str, MergedOrganizationalUnitIdentifier],
     extracted_synopse_contributor_stable_target_ids_by_name: dict[
         str, list[MergedPersonIdentifier]
     ],
     unit_stable_target_ids_by_synonym: dict[str, MergedOrganizationalUnitIdentifier],
     synopse_organization_ids_by_query_string: dict[str, MergedOrganizationIdentifier],
+    synopse_activity: dict[str, Any],
+    contact_merged_id_by_query_string: dict[str, MergedContactPointIdentifier],
 ) -> list[ExtractedActivity]:
     """Transforms Synopse data to extracted activities and load result."""
-    settings = Settings.get()
-    synopse_activity = ActivityMapping.model_validate(
-        load_yaml(settings.synopse.mapping_path / "activity.yaml"),
-    )
     transformed_activities = list(
         transform_synopse_projects_to_mex_activities(
             synopse_projects,
             extracted_primary_source_report_server,
-            unit_stable_target_ids_by_emails,
             extracted_synopse_contributor_stable_target_ids_by_name,
             unit_stable_target_ids_by_synonym,
-            synopse_activity,
+            ActivityMapping.model_validate(synopse_activity),
             synopse_organization_ids_by_query_string,
+            contact_merged_id_by_query_string,
         )
     )
     load(transformed_activities)
@@ -316,7 +285,7 @@ def extracted_synopse_activities(
 
 @asset(group_name="synopse")
 def extracted_synopse_variable_groups(
-    synopse_variables_regular_use_by_thema: dict[str, list[SynopseVariable]],
+    synopse_variables_by_thema: dict[str, list[SynopseVariable]],
     extracted_primary_source_report_server: ExtractedPrimarySource,
     extracted_synopse_resource_stable_target_ids_by_synopse_id: dict[
         str, list[MergedResourceIdentifier]
@@ -325,7 +294,7 @@ def extracted_synopse_variable_groups(
     """Transforms Synopse data to extracted variable groups and load result."""
     transformed_variable_groups = list(
         transform_synopse_variables_to_mex_variable_groups(
-            synopse_variables_regular_use_by_thema,
+            synopse_variables_by_thema,
             extracted_primary_source_report_server,
             extracted_synopse_resource_stable_target_ids_by_synopse_id,
         )
@@ -336,7 +305,7 @@ def extracted_synopse_variable_groups(
 
 @asset(group_name="synopse")
 def extracted_synopse_variables(
-    synopse_variables_regular_use_by_thema: dict[str, list[SynopseVariable]],
+    synopse_variables_by_thema: dict[str, list[SynopseVariable]],
     extracted_primary_source_report_server: ExtractedPrimarySource,
     extracted_synopse_variable_groups: list[ExtractedVariableGroup],
     extracted_synopse_resource_stable_target_ids_by_synopse_id: dict[
@@ -345,7 +314,7 @@ def extracted_synopse_variables(
 ) -> None:
     """Transforms Synopse data to extracted variables and load result."""
     extracted_variables = transform_synopse_variables_to_mex_variables(
-        synopse_variables_regular_use_by_thema,
+        synopse_variables_by_thema,
         extracted_synopse_variable_groups,
         extracted_synopse_resource_stable_target_ids_by_synopse_id,
         extracted_primary_source_report_server,