+++ CHANGELOG.md
@@ -15,12 +15,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - let `load` function accept rule-sets and merged-items, just like sinks
 - let publisher use `load` function with configurable sinks
+- wrap up synopse
 
 ### Deprecated
 
 ### Removed
 
 - remove `PublisherContainer` and `write_merged_items` in favor of generic versions
+- remove `organigram` extractor, it was duplicating `mex.extractors.pipeline.organigram`
 
 ### Fixed